fix(lexarg)!: Move Error, Result to facade crate
These are not core to the parsing process and we can change them as
needed.
