+++ crates/lexarg-error/examples/hello-error.rs
@@ -1,5 +1,4 @@
 use lexarg_error::ErrorContext;
-use lexarg_error::Result;
 
 struct Args {
     thing: String,
@@ -7,7 +6,7 @@ struct Args {
     shout: bool,
 }
 
-fn parse_args() -> Result<Args> {
+fn parse_args() -> Result<Args, String> {
     #![allow(clippy::enum_glob_use)]
     use lexarg_parser::Arg::*;
 
@@ -23,27 +22,36 @@ fn parse_args() -> Result<Args> {
     while let Some(arg) = parser.next_arg() {
         match arg {
             Short("n") | Long("number") => {
-                let value = parser
-                    .next_flag_value()
-                    .ok_or_else(|| ErrorContext::msg("missing required value").within(arg))?;
+                let value = parser.next_flag_value().ok_or_else(|| {
+                    ErrorContext::msg("missing required value")
+                        .within(arg)
+                        .to_string()
+                })?;
                 number = value
                     .to_str()
                     .ok_or_else(|| {
                         ErrorContext::msg("invalid number")
                             .unexpected(Value(value))
                             .within(arg)
+                            .to_string()
                     })?
                     .parse()
-                    .map_err(|e| ErrorContext::msg(e).unexpected(Value(value)).within(arg))?;
+                    .map_err(|e| {
+                        ErrorContext::msg(e)
+                            .unexpected(Value(value))
+                            .within(arg)
+                            .to_string()
+                    })?;
             }
             Long("shout") => {
                 shout = true;
             }
             Value(val) if thing.is_none() => {
-                thing = Some(
-                    val.to_str()
-                        .ok_or_else(|| ErrorContext::msg("invalid string").unexpected(arg))?,
-                );
+                thing = Some(val.to_str().ok_or_else(|| {
+                    ErrorContext::msg("invalid string")
+                        .unexpected(arg)
+                        .to_string()
+                })?);
             }
             Short("h") | Long("help") => {
                 println!("Usage: hello [-n|--number=NUM] [--shout] THING");
@@ -53,21 +61,25 @@ fn parse_args() -> Result<Args> {
                 return Err(ErrorContext::msg("unexpected argument")
                     .unexpected(arg)
                     .within(Value(bin_name))
-                    .into());
+                    .to_string());
             }
         }
     }
 
     Ok(Args {
         thing: thing
-            .ok_or_else(|| ErrorContext::msg("missing argument THING").within(Value(bin_name)))?
+            .ok_or_else(|| {
+                ErrorContext::msg("missing argument THING")
+                    .within(Value(bin_name))
+                    .to_string()
+            })?
             .to_owned(),
         number,
         shout,
     })
 }
 
-fn main() -> Result<()> {
+fn main() -> Result<(), String> {
     let args = parse_args()?;
     let mut message = format!("Hello {}", args.thing);
     if args.shout {
